refactor: 패키지 구조 변경(package-by-feature)
