+++ src/main/java/subway/station/StationRepository.java
@@ -1,4 +1,6 @@
-package subway.domain;
+package subway.station;
+
+import subway.station.Station;
 
 import java.util.ArrayList;
 import java.util.Collections;
